requirements.txt
# requirements.txt
flask>=2.0.0
sqlalchemy>=1.4.0
transformers>=4.0.0
python-dotenv>=0.19.0
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0
torch>=2.0.0
openai>=1.0.0